Rename, cleanup and extract class to own file
