+++ ProjectVersion.py
@@ -1,2 +1,2 @@
 # Format 'V01.09 B01' or 'V01.10'
-ProjectVersionNumber='V01.17 B04'
+ProjectVersionNumber='V01.17 B05'
